Fri Dec 27
I continue with a few problems. The OECH data was removed from the data because the data could not provide us with any data and did not help us. Moreover, the corona data we selected does not clean properly because there is too much empty data or 0 in the cleaning process. This problem needs to be checked and reconsidered. The point I'm stuck on is section  5